:recycle: Refactor: CardList 컴포넌트 모듈화 작업
